Lots of BLIP work
* Implemented server side (BLIPWebSocketListener)
* Added JSON convenience methods
* Fixed Xcode 6.3 compile warnings/errors
* Fixed bugs
* Changed WebSocket default read timeout from 60 sec to infinite

+++ BLIP/BLIPProperties.m
@@ -304,7 +304,7 @@ static const char* kAbbreviations[] = {
 static void appendStr( NSMutableData *data, NSString *str ) {
     const char *utf8 = [str UTF8String];
     size_t size = strlen(utf8)+1;
-    for( unsigned i=0; i<kNAbbreviations; i++ )
+    for( uint8_t i=0; i<kNAbbreviations; i++ )
         if( memcmp(utf8,kAbbreviations[i],size)==0 ) {
             const UInt8 abbrev[2] = {i+1,0};
             [data appendBytes: &abbrev length: 2];
